renamed auth-service to user-service
