+++ .vscode/launch.json
@@ -2,12 +2,12 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "name": "Attach to Auth Service (Docker)",
+      "name": "Attach to User Service (Docker)",
       "type": "node",
       "request": "attach",
       "address": "localhost",
       "port": 9228,
-      "localRoot": "${workspaceFolder}/services/auth-service",
+      "localRoot": "${workspaceFolder}/services/user-service",
       "remoteRoot": "/app",
       "restart": true,
       "autoAttachChildProcesses": true,
