--- services/auth-service/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "ESNext",
-    "outDir": "dist",
-    "rootDir": "src",
-    "moduleResolution": "bundler",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "sourceMap": true,
-},
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
-}
