+++ services/user-service/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "auth-service",
+  "name": "user-service",
   "version": "1.0.0",
   "type": "module",
   "scripts": {
@@ -8,7 +8,7 @@
     "start": "node dist/server.js",
     "clean": "rm -rf dist"
   },
-  "dependencies": {
+    "dependencies": {
     "bcrypt": "6.0.0",
     "dotenv": "^16.5.0",
     "fastify": "^5.3.2",
