+++ services/user-service/src/utils/responseHelpers.ts
